Get checking to behave (I think)
