+++ src/gcages/cmip7_scenariomip/pre_processing.py
@@ -5,649 +5,378 @@ Pre-processing part of the workflow
 from __future__ import annotations
 
 import multiprocessing
-import re
-from collections.abc import Mapping
-from functools import partial
-from typing import TYPE_CHECKING, Callable, Protocol
+from collections import defaultdict
 
+import numpy as np
 import pandas as pd
 from attrs import define
-from pandas_openscm.index_manipulation import update_index_levels_func
+from pandas_openscm.indexing import multi_index_lookup
+from pandas_openscm.parallelisation import ParallelOpConfig, apply_op_parallel_progress
 
 from gcages.assertions import (
     assert_data_is_all_numeric,
     assert_has_index_levels,
     assert_index_is_multiindex,
+    assert_only_working_on_variable_unit_region_variations,
 )
+from gcages.completeness import assert_all_groups_are_complete
 from gcages.exceptions import MissingOptionalDependencyError
-from gcages.renaming import SupportedNamingConventions, convert_variable_name
-from gcages.units_helpers import strip_pint_incompatible_characters_from_units
+from gcages.testing import assert_frame_equal
 
-if TYPE_CHECKING:
-    import pyam
-
-
-def split_variable(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Split the variable into table, gas and sector
-
-    This function is meant for internal use only.
-    It is not subject to the usual set of tests or checks.
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] to work on
-
-    Returns
-    -------
-    :
-        `df` with the variable index level split into table, gas and sector
-    """
-    try:
-        from pandas_indexing.core import extractlevel
-    except ImportError as exc:
-        raise MissingOptionalDependencyError(
-            "split_variable", requirement="pandas_indexing"
-        ) from exc
-
-    res = extractlevel(df, variable="{table}|{gas}|{sector}")
-
-    return res
-
-
-def combine_to_make_variable(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Combine table, gas and sector to make a variable index level
-
-    This function is meant for internal use only.
-    It is not subject to the usual set of tests or checks.
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] to work on
-
-    Returns
-    -------
-    :
-        `df` with the variable index level created from table, gas and sector
-    """
-    try:
-        from pandas_indexing.core import formatlevel
-    except ImportError as exc:
-        raise MissingOptionalDependencyError(
-            "combine_to_make_variable", requirement="pandas_indexing"
-        ) from exc
-
-    res = formatlevel(df, variable="{table}|{gas}|{sector}", drop=True)
-
-    return res
-
-
-def is_ceds_region_sector_level_relevant_variable(
-    v: str,
-    relevant_species: tuple[str, ...] = (
-        "CO2",
-        "CH4",
-        "N2O",
-        "BC",
-        "CO",
-        "NH3",
-        "OC",
-        "NOx",
-        "Sulfur",
-        "VOC",
-    ),
-) -> bool:
-    """
-    Check if a variable is relevant for CEDS region-sector harmonisation
-
-    Parameters
-    ----------
-    v
-        Variable to check
-
-    relevant_species
-        Relevant species for CEDS region-sector harmonisation
-
-        If any of these appear in `v`, we assume that the variable is relevant
-
-    Returns
-    -------
-    :
-        `True` if the variable is relevant for CEDS region-sector harmonisation,
-        otherwise `False`
-    """
-    return any(v_name in v for v_name in relevant_species)
-
-
-def process_transport_variables(  # noqa: PLR0913
-    df: pd.DataFrame,
-    aviation_sector_out: str = "Aircraft",
-    transportation_sector_out: str = "Transportation Sector",
-    aviation_domestic_sector_in: str = "Energy|Demand|Transportation|Domestic Aviation",
-    aviation_international_sector_in: str = "Energy|Demand|Bunkers|International Aviation",  # noqa: E501
-    transportation_sector_in: str = "Energy|Demand|Transportation",
-) -> pd.DataFrame:
-    """
-    Process the transport variables into the convention needed
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] to process
-
-    aviation_sector_out
-        Label to use for the output aviation sector
-
-    transportation_sector_out
-        Label to use for the output transport sector
-
-    aviation_domestic_sector_in
-        Label to assume for the input domestic aviation sector
-
-    aviation_international_sector_in
-        Label to assume for the input international aviation sector
-
-    transportation_sector_in
-        Label to assume for the input transport sector
-
-    Returns
-    -------
-    :
-        `df` with the following changes:
-
-        1. Domestic and international aviation
-           are aggregated to a single aviation sector
-        1. Domestic aviation is subtracted from transportation
-        1. The input transportation sector is dropped
-    """
-    df_split = split_variable(df)
-
-    df_stacked = df_split.unstack("sector").stack("year", future_stack=True)
-
-    df_stacked[aviation_sector_out] = (
-        df_stacked[aviation_domestic_sector_in]
-        + df_stacked[aviation_international_sector_in]
-    )
-
-    df_stacked[transportation_sector_out] = (
-        df_stacked[transportation_sector_in] - df_stacked[aviation_domestic_sector_in]
-    )
-
-    # Should we also consider dropping the aviation split for consistency?
-    keep_sectors = df_stacked.columns.difference([transportation_sector_in])
-    res = combine_to_make_variable(
-        df_stacked[keep_sectors].unstack("year").stack("sector", future_stack=True)
-    )
-
-    return res
-
-
-def aggregate_industry_sector(
-    df: pd.DataFrame,
-    sector_out: str = "Industrial Sector",
-    sectors_to_aggregate: tuple[str, ...] = (
-        "Energy|Demand|Industry",
-        "Energy|Demand|Other Sector",
-        "Industrial Processes",
-        "Other",
-    ),
-) -> pd.DataFrame:
-    """
-    Aggregate the industry sector from its component sectors
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] in which to aggregate the industry sector
-
-    sector_out
-        Name to use for the output sector
-
-    sectors_to_aggregate
-        Sectors to aggregate to create `sector_out`
-
-
-    Returns
-    -------
-    :
-        `df` with the industrial sector included
-    """
-    df_split = split_variable(df)
-
-    df_stacked = df_split.unstack("sector").stack("year", future_stack=True)
-    if "Other" not in df_stacked:
-        # TODO: remove this hack
-        df_stacked["Other"] = 0.0
-
-    df_stacked[sector_out] = df_stacked[list(sectors_to_aggregate)].sum(axis="columns")
-
-    res = combine_to_make_variable(
-        df_stacked.unstack("year").stack("sector", future_stack=True)
-    )
-
-    return res
-
-
-def aggregate_ceds_like_agriculture(
-    df: pd.DataFrame,
-    sector_out: str = "CEDS Agriculture",
-    sectors_to_aggregate: tuple[str, ...] = (
-        "AFOLU|Agriculture",
-        "AFOLU|Land|Land Use and Land-Use Change",
-        "AFOLU|Land|Harvested Wood Products",
-        "AFOLU|Land|Other",
-        "AFOLU|Land|Wetlands",
-    ),
-) -> pd.DataFrame:
-    """
-    Aggregate a CEDS-like agriculture sector from its component sectors
-
-    The default values for `sectors_to_aggregate` are imperfect
-    but the best we have for now.
-    We don't have a better source to harmonise to than CEDS
-    (except for CO2 but ESMs don't use CO2 LULUCF
-    emissions as input anyway, they use land-use change patterns,
-    so this doesn't matter).
-    For SCMs supported by OpenSCM-Runner,
-    this mismatch also doesn't matter as it all gets rolled up to
-    "AFOLU" anyway because SCMs aren't able to handle the difference
-    between e.g. wood harvest and removals due to LULUCF.
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] in which to aggregate the industry sector
-
-    sector_out
-        Name to use for the output sector
-
-    sectors_to_aggregate
-        Sectors to aggregate to create `sector_out`
-
-
-    Returns
-    -------
-    :
-        `df` with the industrial sector included
-    """
-    df_split = split_variable(df)
-
-    df_stacked = df_split.unstack("sector").stack("year", future_stack=True)
-    df_stacked[sector_out] = df_stacked[list(sectors_to_aggregate)].sum(axis="columns")
-
-    res = combine_to_make_variable(
-        df_stacked.unstack("year").stack("sector", future_stack=True)
-    )
-
-    return res
-
-
-DEFAULT_CEDS_RENAMINGS = {
-    "Energy|Supply": "Energy Sector",
-    "Industrial Sector": "Industrial Sector",  # assumes pre-processig
-    "Energy|Demand|Residential and Commercial and AFOFI": "Residential Commercial Other",  # noqa: E501
-    "Product Use": "Solvents Production and Application",
-    "Transportation Sector": "Transportation Sector",
-    "Waste": "Waste",
-    "Aircraft": "Aircraft",
-    "Energy|Demand|Bunkers|International Shipping": "International Shipping",
-    "CEDS Agriculture": "Agriculture",
-    "AFOLU|Agricultural Waste Burning": "Agricultural Waste Burning",
-    "AFOLU|Land|Fires|Forest Burning": "Forest Burning",
-    "AFOLU|Land|Fires|Grassland Burning": "Grassland Burning",
-    "AFOLU|Land|Fires|Peat Burning": "Peat Burning",
-}
+REQUIRED_GRIDDING_SPECIES_IAMC: tuple[str, ...] = (
+    "CO2",
+    "CH4",
+    "N2O",
+    "BC",
+    "CO",
+    "NH3",
+    "OC",
+    "NOx",
+    "Sulfur",
+    "VOC",
+)
 """
-Default renamigns used by [rename_and_cut_to_ceds_aligned_sectors][]
+Species to prepare for gridding (in the IAMC naming convention)
+
+All species in this tuple must be provided in the input.
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+INDUSTRIAL_SECTOR_CEDS_COMPONENTS_IAMC: tuple[str, ...] = (
+    "Energy|Demand|Industry",
+    "Energy|Demand|Other Sector",
+    "Industrial Processes",
+    "Other",
+)
+"""
+Sectors (in the IAMC naming convention) that make up the CEDS industrial sector
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+DOMESTIC_AVIATION_SECTOR_IAMC: str = "Energy|Demand|Transportation|Domestic Aviation"
+"""
+Assumed sector for the domestic aviation sector (IAMC naming convention)
+"""
+
+INTERNATIONAL_AVIATION_SECTOR_IAMC: str = "Energy|Demand|Bunkers|International Aviation"
+"""
+Assumed reporting for the international aviation sector (IAMC naming convention)
+"""
+
+TRANSPORTATION_SECTOR_IAMC: str = "Energy|Demand|Transportation"
+"""
+Assumed reporting for the transportation sector (IAMC naming convention)
+"""
+
+AGRICULTURE_SECTOR_COMPONENTS_IAMC: tuple[str, ...] = (
+    "AFOLU|Agriculture",
+    "AFOLU|Land|Land Use and Land-Use Change",
+    "AFOLU|Land|Harvested Wood Products",
+    "AFOLU|Land|Other",
+    "AFOLU|Land|Wetlands",
+)
+"""
+Sectors that make up the CEDS agriculture sector (IAMC naming convention)
+
+Not all of these components are required.
+If they are not supplied, they are assumed to be zero.
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_GRIDDING_SECTORS_REGIONAL_IAMC: tuple[str, ...] = (
+    "Energy|Supply",
+    *INDUSTRIAL_SECTOR_CEDS_COMPONENTS_IAMC,
+    "Energy|Demand|Residential and Commercial and AFOFI",
+    "Product Use",
+    # Technically, domestic aviation could be reported just
+    # at the world level and it would be fine.
+    # In practice, no-one does that and the logic is much simpler
+    # if we assume it has to be reported regionally
+    # (because then domestic aviation and transport are on the same regional 'grid')
+    # so do that for now.
+    DOMESTIC_AVIATION_SECTOR_IAMC,
+    TRANSPORTATION_SECTOR_IAMC,
+    # The rest of TRANSPORTATION_SECTOR_REQUIRED_REPORTING_IAMC
+    # can be reported at the world level only and it is fine
+    "Waste",
+    # Note: AFOLU|Agriculture is the only compulsory component for agriculture
+    # which is why we don't use
+    # *AGRICULTURE_SECTOR_COMPONENTS_IAMC
+    "AFOLU|Agriculture",
+    "AFOLU|Agricultural Waste Burning",
+    "AFOLU|Land|Fires|Forest Burning",
+    "AFOLU|Land|Fires|Grassland Burning",
+    "AFOLU|Land|Fires|Peat Burning",
+)
+"""
+Sectors (IAMC naming) required in the input at the regional level for gridding
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_GRIDDING_SECTORS_WORLD_IAMC: tuple[str, ...] = (
+    INTERNATIONAL_AVIATION_SECTOR_IAMC,
+    "Energy|Demand|Bunkers|International Shipping",
+)
+"""
+Sectors (IAMC naming) required in the input at the world level for gridding
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_REGIONAL_VARIABLES_IAMC: tuple[str, ...] = tuple(
+    f"Emissions|{species}|{sector}"
+    for species in REQUIRED_GRIDDING_SPECIES_IAMC
+    for sector in REQUIRED_GRIDDING_SECTORS_REGIONAL_IAMC
+)
+"""
+Variables required at the regional level (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_WORLD_VARIABLES_IAMC: tuple[str, ...] = tuple(
+    f"Emissions|{species}|{sector}"
+    for species in REQUIRED_GRIDDING_SPECIES_IAMC
+    for sector in REQUIRED_GRIDDING_SECTORS_WORLD_IAMC
+)
+"""
+Variables required at the world level (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_REGIONAL_INDEX_IAMC = pd.MultiIndex(
+    names=["variable"],
+    levels=[REQUIRED_REGIONAL_VARIABLES_IAMC],
+    codes=[np.arange(len(REQUIRED_REGIONAL_VARIABLES_IAMC))],
+)
+"""
+The required index for regional data for gridding (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_REGIONAL_VARIABLES_IAMC: tuple[str, ...] = tuple(
+    f"Emissions|{species}|{sector}"
+    for species in REQUIRED_GRIDDING_SPECIES_IAMC
+    for sector in REQUIRED_GRIDDING_SECTORS_REGIONAL_IAMC
+)
+"""
+Variables required at the regional level (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+ALL_RELEVANT_REGIONAL_VARIABLES_IAMC = tuple(
+    # Use of set avoids double counting
+    {
+        *REQUIRED_REGIONAL_VARIABLES_IAMC,
+        *(
+            f"Emissions|{species}|{sector}"
+            for species in REQUIRED_GRIDDING_SPECIES_IAMC
+            for sector in AGRICULTURE_SECTOR_COMPONENTS_IAMC
+        ),
+    }
+)
+"""
+Variables that are relevant at the regional level (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+ALL_RELEVANT_REGIONAL_INDEX_IAMC = pd.MultiIndex(
+    names=["variable"],
+    levels=[ALL_RELEVANT_REGIONAL_VARIABLES_IAMC],
+    codes=[np.arange(len(ALL_RELEVANT_REGIONAL_VARIABLES_IAMC))],
+)
+
+"""
+The index of all relevant levels for regional data for gridding (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+REQUIRED_WORLD_INDEX_IAMC = pd.MultiIndex(
+    names=["variable"],
+    levels=[REQUIRED_WORLD_VARIABLES_IAMC],
+    codes=[np.arange(len(REQUIRED_WORLD_VARIABLES_IAMC))],
+)
+"""
+The required index for world data for gridding (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+ALL_RELEVANT_WORLD_VARIABLES_IAMC = REQUIRED_WORLD_VARIABLES_IAMC
+"""
+Variables that are relevant at the world level (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
+"""
+
+ALL_RELEVANT_WORLD_INDEX_IAMC = REQUIRED_WORLD_INDEX_IAMC
+"""
+The index of all relevant levels for world data for gridding (IAMC naming convention)
+
+This variable is provided as a global variable for clarity and consistency.
+If you change it, we do not guarantee the package's performance.
 """
 
 
-def rename_and_cut_to_ceds_aligned_sectors(
-    df: pd.DataFrame,
-    renamings: Mapping[str, str] | None = None,
-) -> pd.DataFrame:
+@define
+class SplitData:
     """
-    Rename sectors to CEDS sectors and only keep the renamed sectors
-
-    Parameters
-    ----------
-    df
-        [pd.DataFrame][pandas.DataFrame] to work on
-
-    renamings
-        Renamings to apply
-
-        If not supplied, we use [DEFAULT_CEDS_RENAMINGS]
-
-    Returns
-    -------
-    :
-        `df` with only renamed sectors
+    Container for holding data split into regional and world data
     """
-    if renamings is None:
-        renamings = DEFAULT_CEDS_RENAMINGS
 
-    df_stacked = split_variable(df).unstack("sector").stack("year", future_stack=True)
+    world: pd.DataFrame
+    """Data at the world level"""
 
-    # TODO: remove this
-    if "AFOLU|Land|Fires|Peat Burning" not in df_stacked:
-        df_stacked["AFOLU|Land|Fires|Peat Burning"] = 0.0
-
-    renamed = df_stacked.rename(renamings, axis="columns", errors="raise")
-
-    res = combine_to_make_variable(
-        renamed[renamings.values()].unstack("year").stack("sector", future_stack=True)
-    )
-
-    return res
+    regional: pd.DataFrame
+    """Data at the regional level"""
 
 
-def create_global_workflow_input_from_region_sector_input(  # noqa: PLR0913
-    region_sector_df: pd.DataFrame,
-    level_separator: str = "|",
-    n_separators_for_single_sector: int = 2,
-    world_region: str = "World",
-    co2_fossil_sectors: tuple[str, ...] = (
-        "Energy Sector",
-        "Industrial Sector",
-        "Residential Commercial Other",
-        "Solvents Production and Application",
-        "Transportation Sector",
-        "Waste",
-        "Aircraft",
-        "International Shipping",
-    ),
-    co2_fossil_out_name: str = "Emissions|CO2|Energy and Industrial Processes",
-    co2_afolu_out_name: str = "Emissions|CO2|AFOLU",
-) -> pd.DataFrame:
-    """
-    Create input for the global workflow
+def split_world_and_regional_data(indf: pd.DataFrame, world_region: str) -> SplitData:
+    world_idxer = indf.index.get_level_values("region") == world_region
+    world_data = indf.loc[world_idxer]
+    regional_data = indf.loc[~world_idxer]
 
-    Parameters
-    ----------
-    region_sector_df
-        [pd.DataFrame][pandas.DataFrame] with region-sector harmonisation information
+    return SplitData(world=world_data, regional=regional_data)
 
-    level_separator
-        Separator used between levels in variable strings
 
-    n_separators_for_single_sector
-        Number of separators in a variable name when there is a single sector
-
-    world_region
-        Name of the world/world total region
-
-    co2_fossil_sectors
-        Sectors whose emissions originate from fossil CO2
-
-    co2_fossil_out_name
-        Name of the CO2 fossil-reservoir emissions output variable
-
-    co2_afolu_out_name
-        Name of the CO2 AFOLU-reservoir (i.e. biosphere) emissions output variable
-
-    Returns
-    -------
-    :
-        Output that can be used with the global-based workflow
-    """
-    region_sector_df_variables = region_sector_df.index.get_level_values(
-        "variable"
-    ).unique()
-    not_single_sector = region_sector_df_variables[
-        region_sector_df_variables.str.count(re.escape(level_separator))
-        != n_separators_for_single_sector
+def assert_no_obvious_double_counting(sum_over_values: list[str]) -> None:
+    double_counting = [
+        s
+        for s in sum_over_values
+        # If the value is a subset of another value,
+        # we will double count
+        if any(f"{s}|" in s_other for s_other in sum_over_values)
     ]
-    if not not_single_sector.empty:
-        msg = (
-            "The following variables in `region_sector_df` don't have a single sector"
-            f"{not_single_sector}"
-        )
+    if double_counting:
+        msg = f"Likely double counting {double_counting=} {sum_over_values=}"
+
         raise AssertionError(msg)
 
-    if "World" in region_sector_df.index.get_level_values("region"):
-        raise AssertionError
 
-    def get_sum_over_sectors_and_regions(idf: pd.DataFrame) -> pd.DataFrame:
-        return (
-            update_index_levels_func(
-                idf,
-                dict(
-                    variable=lambda x: level_separator.join(
-                        x.split(level_separator)[:n_separators_for_single_sector]
-                    ),
-                    region=lambda x: world_region,
-                ),
-            )
-            .groupby(idf.index.names)
-            .sum()
-        )
+def get_region_sector_totals(
+    indf: pd.DataFrame, world_region: str, region_level: str
+) -> pd.DataFrame:
+    try:
+        from pandas_indexing.core import assignlevel, extractlevel, formatlevel
+    except ImportError as exc:
+        raise MissingOptionalDependencyError(
+            "get_region_sector_totals", requirement="pandas_indexing"
+        ) from exc
 
-    region_sector_world_totals = get_sum_over_sectors_and_regions(region_sector_df)
-    non_co2 = region_sector_world_totals.loc[
-        region_sector_world_totals.index.get_level_values("variable") != "Emissions|CO2"
-    ]
+    indf_split = extractlevel(indf, variable="{table}|{species}|{sectors}")
 
-    region_sector_df_variable_level = region_sector_df.index.get_level_values(
-        "variable"
+    assert_no_obvious_double_counting(
+        indf_split.index.get_level_values("sectors").unique().tolist()
     )
-    co2_idx = region_sector_df_variable_level.str.startswith("Emissions|CO2")
-    fossil_sector_idx = region_sector_df_variable_level.map(
-        lambda x: any(x.endswith(s) for s in co2_fossil_sectors)
+    assert_no_obvious_double_counting(
+        indf_split.index.get_level_values("region").unique().tolist()
     )
-
-    co2_fossil_components = region_sector_df.loc[co2_idx & fossil_sector_idx]
-    co2_fossil = update_index_levels_func(
-        get_sum_over_sectors_and_regions(co2_fossil_components),
-        dict(variable=lambda x: co2_fossil_out_name),
-    )
-
-    # Doing it like this means we get all the burning in here too.
-    # That is probably double counting,
-    # although simple climate models don't have interactive fire emissions
-    # so maybe it is fine.
-    co2_afolu_components = region_sector_df.loc[co2_idx & ~fossil_sector_idx]
-    co2_afolu = update_index_levels_func(
-        get_sum_over_sectors_and_regions(co2_afolu_components),
-        dict(variable=lambda x: co2_afolu_out_name),
-    )
-
-    res = pd.concat(
-        [
-            v.reorder_levels(co2_afolu.index.names)
-            for v in [co2_fossil, co2_afolu, non_co2]
-        ]
+    res = formatlevel(
+        assignlevel(
+            indf_split.groupby(
+                indf_split.index.names.difference([region_level, "sectors"])
+            ).sum(),
+            region=world_region,
+        ),
+        variable="{table}|{species}",
+        drop=True,
     )
 
     return res
 
 
-def create_global_workflow_input_from_raw_input(df: pd.DataFrame) -> pd.DataFrame:
+def get_raw_data_for_gridding(indf: pd.DataFrame, world_region: str) -> pd.DataFrame:
+    # hi
+    pass
     """
-    Create global workflow input from raw input
+    You'll likely want to call
 
-    This should only be passed input that is not created
-    from the region-sector relevant input
-
-    Parameters
-    ----------
-    df
-        Raw input
-
-    Returns
-    -------
-    :
-        Global workflow input derived from `df`
+    [assert_data_is_compatible_with_pre_processing][(m).]
+    before using this as it gives better error messages if you have missing data.
     """
-    # For now, we can't use these.
-    # If we figure out downscaling, we could
-    # (but likely to be model-dependent hence hard to include here).
-    out = strip_pint_incompatible_characters_from_units(
-        df.loc[~df.index.get_level_values("unit").str.contains("equiv")],
-        units_index_level="unit",
+    split_data = split_world_and_regional_data(indf, world_region=world_region)
+
+    regional_data = multi_index_lookup(
+        split_data.regional, ALL_RELEVANT_REGIONAL_INDEX_IAMC
     )
+    world_data = multi_index_lookup(split_data.world, ALL_RELEVANT_WORLD_INDEX_IAMC)
 
-    return out
+    res = pd.concat([world_data, regional_data.reorder_levels(world_data.index.names)])
+
+    return res
 
 
-class InternalConsistencyError(ValueError):
-    """
-    Raised when there is an internal consistency issue
-    """
+def drop_out_domestic_aviation(indf: pd.DataFrame) -> pd.DataFrame:
+    if TRANSPORTATION_SECTOR_IAMC not in DOMESTIC_AVIATION_SECTOR_IAMC:
+        msg = "Logic of this is broken"
+        raise AssertionError(msg)
 
-    def __init__(
-        self,
-        indf: pd.DataFrame,
-        reporting_issues: pd.DataFrame,
-        dsd: DataStructureDefinitionLike,
-        level_separator: str,
-    ) -> None:
-        reporting_issues_wide = (
-            reporting_issues.melt(ignore_index=False, var_name="aggregation")
-            .set_index("aggregation", append=True)["value"]
-            .unstack("year")
+    res = indf.loc[
+        ~indf.index.get_level_values("variable").str.endswith(
+            DOMESTIC_AVIATION_SECTOR_IAMC
         )
-        indf_variables = indf.index.get_level_values("variable").unique()
+    ]
 
-        msg_l = [
-            "There are reporting issues in your data.",
-            "The issue occurs for the following variables:",
-        ]
-        for i, variable_with_issue in enumerate(
-            reporting_issues.index.get_level_values("variable").unique()
-        ):
-            msg_l.extend(("", f"{i + 1}. {variable_with_issue}"))
-
-            dsd_components = dsd.variable[variable_with_issue].components
-            if dsd_components is not None:
-                msg_l.append(
-                    "  - these are the expected component variables "
-                    "from the data structure definition: "
-                    f"{sorted(dsd_components)}",
-                )
-                msg_l.append(
-                    "  - if you have reported data "
-                    "that is not part of the expected component variables, "
-                    "this may be the issue",
-                )
-                included_in_df = sorted(
-                    indf_variables.intersection(dsd_components).tolist()
-                )
-                missing_from_df = sorted(
-                    list(set(dsd_components) - set(indf_variables))
-                )
-
-                msg_l.extend(
-                    (
-                        "  - of the given components, these variables:",
-                        f"    - are included in the data: {included_in_df}",
-                        f"    - are missing from the data: {missing_from_df}",
-                    )
-                )
-
-            else:
-                indf_variables_in_hierarchy = [
-                    v
-                    for v in indf_variables
-                    if v.startswith(variable_with_issue)
-                    and v.split(variable_with_issue)[-1].count(level_separator) == 1
-                ]
-                msg_l.append(
-                    "  - the variable has no components "
-                    "in the data structure definition, "
-                    "so we assume that we are insted checking "
-                    "over the variables in the hierarchy: "
-                    f"{indf_variables_in_hierarchy}"
-                )
-
-        msg_l.extend(
-            (
-                "",
-                "Here is a view of the difference between the variable values",
-                "and the sum of their expected components:",
-                f"{reporting_issues_wide}",
-            )
-        )
-        error_msg = "\n".join(msg_l)
-        super().__init__(error_msg)
+    return res
 
 
-class DataStructureDefinitionLike(Protocol):
-    """
-    Object that behaves like the nomenclature package's DataStructureDefinition class
-    """
-
-    def check_aggregate(self, df: pyam.IamDataFrame) -> pd.DataFrame | None:
-        """
-        Check aggregate along the variable hierarchy
-
-        Parameters
-        ----------
-        df
-            Data to check
-
-        Returns
-        -------
-        :
-            List of errors if there are any, otherwise `None`
-        """
-
-
-def assert_data_is_internally_consistent(
-    indf: pd.DataFrame,
-    dsd: DataStructureDefinitionLike,
-    rtol: float = 1e-3,
-    atol: float = 1e-8,
-    level_separator: str = "|",
+def assert_data_is_compatible_with_pre_processing(
+    indf: pd.DataFrame, world_region: str, region_level: str
 ) -> None:
-    """
-    Assert that data is internally consistent
+    indf_drop_all_nan_cols = indf.dropna(how="all", axis="columns")
+    if indf_drop_all_nan_cols.isnull().any().any():
+        msg = f"NaNs after dropping unreported times:\n{indf_drop_all_nan_cols}"
+        raise AssertionError(msg)
 
-    Parameters
-    ----------
-    indf
-        Data to check
+    assert_only_working_on_variable_unit_region_variations(indf)
 
-    dsd
-        Data structure definition to use for checking
+    split_data = split_world_and_regional_data(indf, world_region=world_region)
 
-    rtol
-        Relative tolerance to apply while checking consistency
-
-    atol
-        Absolute tolerance to apply while checking consistency
-
-    level_separator
-        Separator between levels in the variable hierarchy
-
-    Raises
-    ------
-    InternalConsistencyError
-        The data is not internally consistent
-
-    """
-    try:
-        import pyam
-    except ImportError as exc:
-        raise MissingOptionalDependencyError(
-            "assert_data_is_internally_consistent", requirement="pyam"
-        ) from exc
-
-    reporting_issues = dsd.check_aggregate(
-        pyam.IamDataFrame(indf), rtol=rtol, atol=atol
+    assert_all_groups_are_complete(
+        split_data.world, complete_index=REQUIRED_WORLD_INDEX_IAMC
     )
-    if reporting_issues is None:
-        return
-
-    raise InternalConsistencyError(
-        indf=indf,
-        reporting_issues=reporting_issues,
-        dsd=dsd,
-        level_separator=level_separator,
+    assert_all_groups_are_complete(
+        split_data.regional, complete_index=REQUIRED_REGIONAL_INDEX_IAMC
     )
+
+    data_for_gridding_raw = get_raw_data_for_gridding(indf, world_region=world_region)
+
+    # Drop out domestic aviation because it is a subsector of transport
+    data_for_gridding_raw_to_sum = drop_out_domestic_aviation(data_for_gridding_raw)
+
+    data_for_gridding_totals = get_region_sector_totals(
+        data_for_gridding_raw_to_sum,
+        world_region=world_region,
+        region_level=region_level,
+    )
+
+    indf_global_totals = multi_index_lookup(
+        split_data.world, data_for_gridding_totals.index
+    )
+
+    assert_frame_equal(data_for_gridding_totals, indf_global_totals)
 
 
 @define
@@ -655,7 +384,7 @@ class CMIP7ScenarioMIPPreProcessingResult:
     """
     Result of pre-processing with [CMIP7ScenarioMIPPreProcessor][]
 
-    This is more complex than normal,
+    This has more components than normal,
     because we need to support both the 'normal' global path
     and harmonising at the region-sector level.
     """
@@ -665,92 +394,44 @@ class CMIP7ScenarioMIPPreProcessingResult:
     Emissions that can be used with the 'normal' global workflow
     """
 
-    global_workflow_emissions_gcages: pd.DataFrame
+    global_workflow_emissions_iamc: pd.DataFrame
     """
-    Emissions that can be used with the 'normal' global workflow using gcages naming
+    Emissions consistent with those that can be used with the 'normal' global workflow
+
+    The difference is that these are reported with IAMC naming,
+    which isn't compatible with our runners
+    (so probably not what you want to use,
+    but perhaps helpful for plotting and direct comparisons).
     """
 
-    region_sector_workflow_emissions: pd.DataFrame
+    gridding_workflow_emissions: pd.DataFrame
     """
-    Emissions that can be used with the 'normal' global workflow
+    Emissions that can be used with the gridding workflow
     """
 
-    reaggregated_emissions: pd.DataFrame
-    """
-    Emissions that have been reaggregated, but otherwise not altered from the input
-    """
+
+def do_pre_processing(
+    indf: pd.DataFrame, world_region: str
+) -> CMIP7ScenarioMIPPreProcessingResult:
+    split_data = split_world_and_regional_data(indf, world_region=world_region)
+    breakpoint()
 
 
 @define
 class CMIP7ScenarioMIPPreProcessor:
     """
     Pre-processor for CMIP7's ScenarioMIP
-    """
 
-    is_region_sector_relevant_variable: Callable[[str], bool] = (
-        is_ceds_region_sector_level_relevant_variable
-    )
-    """
-    Function to use to determine
-    whether a variable is relevant for region-sector harmonisation or not
-    """
-    reprocess_transport_variables: Callable[[pd.DataFrame], pd.DataFrame] = (
-        process_transport_variables
-    )
-    """
-    Function to use to re-process the transport variables
-    """
-
-    calculate_industrial_sector: Callable[[pd.DataFrame], pd.DataFrame] = (
-        aggregate_industry_sector
-    )
-    """
-    Function to use to calculate the industrial sector
-    """
-
-    calculate_ceds_like_agriculture_sector: Callable[[pd.DataFrame], pd.DataFrame] = (
-        aggregate_ceds_like_agriculture
-    )
-    """
-    Function to use to calculate a CEDS-like agriculture sector
-    """
-
-    convert_to_and_extract_ceds_sectors: Callable[[pd.DataFrame], pd.DataFrame] = (
-        rename_and_cut_to_ceds_aligned_sectors
-    )
-    """
-    Function to use to convert to CEDS sectors and extract only CEDS sectors
-
-    Called after reaggregation has already been done
-    """
-
-    convert_region_sector_to_global_workflow_input: Callable[
-        [pd.DataFrame], pd.DataFrame
-    ] = create_global_workflow_input_from_region_sector_input
-    """
-    Function to use to get the global workflow input that comes from region-sector input
-
-    Called after processing to region-sector input has been done
-    """
-
-    convert_raw_to_global_workflow_input: Callable[[pd.DataFrame], pd.DataFrame] = (
-        create_global_workflow_input_from_raw_input
-    )
-    """
-    Function to use to get the global workflow input that comes from raw input
-
-    This is applied on the input emissions except for the bits
-    which are deemed relevant for region-sector level stuff.
+    Like most pre-processing,
+    this is quite locked up and context-specific.
+    The logic also goes through multiple layers,
+    which makes it hard to clearly identify
+    how techniques like dependency injection could actually be used.
     """
 
     world_region: str = "World"
     """
     String that identifies the world (i.e. global total) region
-    """
-
-    data_structure_definition: DataStructureDefinitionLike | None = None
-    """
-    Data structure definition
     """
 
     run_checks: bool = True
@@ -794,80 +475,64 @@ class CMIP7ScenarioMIPPreProcessor:
         if self.run_checks:
             assert_index_is_multiindex(in_emissions)
             assert_data_is_all_numeric(in_emissions)
-            assert_has_index_levels(in_emissions, ["variable", "unit"])
-
-            if self.data_structure_definition is None:
-                msg = (
-                    "`self.data_structure_definition` must not be None "
-                    "to run the checks"
-                )
-                raise TypeError(msg)
-
-            assert_data_is_internally_consistent(
-                in_emissions, self.data_structure_definition
+            assert_has_index_levels(
+                in_emissions, ["variable", "unit", "model", "scenario", "region"]
             )
 
-        if in_emissions.columns.name != "year":
-            # Make later processing easier without annoying users
-            in_emissions = in_emissions.copy()
-            in_emissions.columns.name = "year"
-
-        world_data_idx = (
-            in_emissions.index.get_level_values("region") == self.world_region
-        )
-        region_sector_relevant_variable_idx = in_emissions.index.get_level_values(
-            "variable"
-        ).map(self.is_region_sector_relevant_variable)
-
-        in_emissions_region_sector_relevant = in_emissions.loc[
-            region_sector_relevant_variable_idx & ~world_data_idx
-        ]
-        in_emissions_not_region_sector_relevant = in_emissions.loc[
-            ~region_sector_relevant_variable_idx & world_data_idx
-        ]
-
-        reaggregated_emissions = self.calculate_ceds_like_agriculture_sector(
-            self.calculate_industrial_sector(
-                self.reprocess_transport_variables(in_emissions_region_sector_relevant)
+            assert_data_is_compatible_with_pre_processing(
+                in_emissions, world_region=self.world_region, region_level="region"
             )
-        )
 
-        region_sector_workflow_emissions = self.convert_to_and_extract_ceds_sectors(
-            reaggregated_emissions
-        )
+            if in_emissions.columns.name != "year":
+                msg = "The input emissions' column name should be 'year'"
+                raise AssertionError(msg)
 
-        global_workflow_emissions_from_region_sector = (
-            self.convert_region_sector_to_global_workflow_input(
-                region_sector_df=region_sector_workflow_emissions
+        res_g = apply_op_parallel_progress(
+            func_to_call=do_pre_processing,
+            world_region=self.world_region,
+            iterable_input=(
+                gdf for _, gdf in in_emissions.groupby(["model", "scenario"])
+            ),
+            parallel_op_config=ParallelOpConfig.from_user_facing(
+                progress=self.progress,
+                max_workers=self.n_processes,
+            ),
+        )
+        res_d = defaultdict(list)
+        for res_ms in res_g:
+            if self.run_checks:
+                for attr in [
+                    "global_workflow_emissions",
+                    "gridding_workflow_emissions",
+                ]:
+                    if res_ms.isnull().any().any():
+                        ms = (
+                            res_ms.index.droplevels(
+                                res_ms.index.names.difference(["model", "scenario"])
+                            )
+                            .drop_duplicates()
+                            .tolist()
+                        )
+                        msg = f"NaNs in res.{attr} for {ms}"
+                        raise AssertionError(msg)
+
+            for attr in [
+                "global_workflow_emissions",
+                "global_workflow_emissions_iamc",
+                "gridding_workflow_emissions",
+            ]:
+                res_d[attr].append(getattr(res_ms, attr))
+
+        res = CMIP7ScenarioMIPPreProcessingResult(
+            **{k: pd.concat(v) for k, v in res_d.items()}
+        )
+        if self.run_checks:
+            assert_all_expected_variables_and_regions_included(
+                res.gridding_workflow_emissions
             )
-        )
-        global_workflow_emissions_from_raw = self.convert_raw_to_global_workflow_input(
-            df=in_emissions_not_region_sector_relevant
-        )
+            assert_gridding_and_global_emissions_are_consistent(
+                gridding_workflow_emissions=res.gridding_workflow_emissions,
+                global_workflow_emissions_iamc=res.global_workflow_emissions_iamc,
+            )
 
-        global_workflow_emissions = pd.concat(
-            [
-                global_workflow_emissions_from_region_sector,
-                global_workflow_emissions_from_raw.reorder_levels(
-                    global_workflow_emissions_from_region_sector.index.names
-                ),
-            ]
-        )
-
-        global_workflow_emissions_gcages = update_index_levels_func(
-            global_workflow_emissions,
-            {
-                "variable": partial(
-                    convert_variable_name,
-                    from_convention=SupportedNamingConventions.CMIP7_SCENARIOMIP,
-                    to_convention=SupportedNamingConventions.GCAGES,
-                )
-            },
-        )
-
-        return CMIP7ScenarioMIPPreProcessingResult(
-            global_workflow_emissions=global_workflow_emissions,
-            global_workflow_emissions_gcages=global_workflow_emissions_gcages,
-            region_sector_workflow_emissions=region_sector_workflow_emissions,
-            reaggregated_emissions=reaggregated_emissions,
-        )
+        return res
